pyproject.toml
[project]
name = "mindmotion"
version = "0.1.0"
authors = [
    {name = "Denis Tokarev",email = "wf.denis.tokarev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mypy (>=1.15.0,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]  # poetry add ... | poetry install


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.education_platform.dependencies]   # poetry add ... --group education_platform | poetry install --with=education_platform
django = "^5.2"
python-dotenv = "^1.1.0"
django-stubs = "^5.1.3"
django-debug-toolbar = "^5.1.0"
